fix: PluginManagerTest
