+++ test/src/test/resources/plugins/security3037-update-center.json
